+++ Wireframe.module.php
@@ -8,7 +8,7 @@ namespace ProcessWire;
  * Wireframe is an output framework with MVC inspired architecture for ProcessWire CMS/CMF.
  * See README.md or https://wireframe-framework.com for more details.
  *
- * @version 0.7.0
+ * @version 0.8.0
  * @author Teppo Koivula <teppo@wireframe-framework.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -73,8 +73,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Create directories automatically?
      *
-     * This property is only used by the module configuration screen. Contains an array of
-     * directories to create automatically.
+     * Used by the module configuration screen. Contains an array of directories that should be automatically created.
      *
      * @var array
      */
@@ -110,9 +109,12 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Keep track of whether Wireframe has already been initialized
      *
-     * @var bool
+     * This information is stored in an array for it to work properly with multi-instance support;
+     * in such cases we need to make sure that initOnce() is run once per ProcessWire instance.
+     *
+     * @var array
      */
-    protected $initialized = false;
+    protected static $initialized = [];
 
     /**
      * Initialize Wireframe
@@ -159,10 +161,10 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      *
      * @return bool True on first run, false if already initialized.
      */
-    protected function initOnce(): bool {
+    public function initOnce(): bool {
 
         // bail out early if already initialized
-        if ($this->initialized) return false;
+        if (static::isInitialized($this->wire()->instanceID)) return false;
 
         // set config settings
         $this->setConfig();
@@ -180,11 +182,27 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
         $this->addHooks();
 
         // remember that this method has been run and return true
-        $this->initialized = true;
+        static::$initialized[] = $this->wire()->instanceID;
 
         // return true on first run
         return true;
 
+    }
+
+    /**
+     * Check if Wireframe has already been initialized
+     *
+     * As long as $instanceID is provided, this method will work on a multi-instance ProcessWire
+     * setup. If $instanceID is left out (null), we get it from the wire() function instead.
+     *
+     * @param null|int $instanceID ProcessWire instance ID. This parameter is optional but recommended.
+     * @return bool True if initialized, false if not.
+     */
+    public static function isInitialized(int $instanceID = null): bool {
+        return in_array(
+            is_null($instanceID) ? wire()->instanceID : $instanceID,
+            static::$initialized
+        );
     }
 
     /**
@@ -195,8 +213,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      */
     public function ___setConfig(array $config = []): Wireframe {
 
-        // default config settings; if you need to customize or override any of
-        // these, copy this array to /site/config.php as $config->wireframe
+        // default config settings; if you need to customize or override any of these, copy this array to
+        // your site config file (/site/config.php) as $config->wireframe
         $config_defaults = [
             'include_paths' => [
                 // '/path/to/shared/libraries/',
@@ -221,6 +239,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
                 'views' => $this->wire('config')->paths->templates . "views/",
                 'layouts' => $this->wire('config')->paths->templates . "layouts/",
                 'partials' => $this->wire('config')->paths->templates . "partials/",
+                'components' => $this->wire('config')->paths->templates . "components/",
                 'controllers' => $this->wire('config')->paths->templates . "controllers/",
             ],
             'urls' => [
@@ -278,18 +297,17 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Add Wireframe namespaces to ProcessWire's class autoloader
      *
-     * This method makes ProcessWire's class autoloader aware of the Wireframe namespaces, which
-     * enables us to instantiate – or call static methods from – Wireframe objects without first
-     * requiring the related PHP file.
+     * This method makes ProcessWire's class autoloader aware of the Wireframe namespaces, which enables us to
+     * instantiate – or call static methods from – Wireframe objects without first requiring the related PHP file.
      *
-     * If you need to add additional namespaces (or additional paths for namespaces added here),
-     * access the $classLoader API variable directly from your own code. If you want to override
-     * these definitions, you should first call $classLoader->removeNamespace($namespace, $path)
-     * – and then re-add the same namespace with your own path.
+     * If you want to add additional namespaces (or additional paths for the namespaces added here), you can access
+     * the $classLoader API variable directly via your own code. If you want to override namespaces added here, you
+     * should call $classLoader->removeNamespace($namespace, $path) before re-adding the namespace with a new path.
      */
     protected function addNamespaces() {
         $namespaces = [
             'Wireframe' => $this->wire('config')->paths->Wireframe . 'lib/',
+            'Wireframe\Component' => $this->paths->components,
             'Wireframe\Controller' => $this->paths->controllers,
             'Wireframe\Lib' => $this->paths->lib,
         ];
@@ -326,6 +344,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
      *
      * @see pageLayout() for the Page::layout(), Page::getLayout(), and Page::setLayout() implementation.
      * @see pageView() for the Page::view(), Page::getView(), and Page::setView() implementation.
+     *
+     * @todo Page::getViewTemplate + Page::setViewTemplate
      */
     protected function addHooks() {
 
@@ -344,8 +364,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Check if a redirect should occur
      *
-     * Look for redirect fields within config settings. If present, check if the
-     * page has a value in one of those and if a redirect should be performed.
+     * Look for redirect fields within config settings. If present, check if the page has a value in one of those and
+     * if a redirect should be performed.
      */
     public function ___checkRedirects() {
 
@@ -439,8 +459,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Initialization method for the Controller
      *
-     * Controller is optional component in Wireframe, but if a Controller file is found, we'll
-     * attempt to instantiate an object from it.
+     * Controller is optional component in Wireframe, but if a Controller file is found, we'll attempt to instantiate
+     * an object from it.
      *
      * @return \Wireframe\Controller|null Controller object or null.
      *
@@ -469,8 +489,10 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Set current view
      *
-     * Default value is 'default', but view() method of the $page object or GET param
-     * 'view' (if configured) can be used to override the default value.
+     * Default value is 'default', but the setView() method of the $page object or GET param 'view' (if configured so)
+     * can be used to override the default value.
+     *
+     * @todo Page::getViewTemplate()
      */
     public function ___setView() {
 
@@ -628,6 +650,9 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
             $event->return = $event->object->_wireframe_view;
         } else {
             $event->object->_wireframe_view = $event->arguments[0] ?? '';
+            $event->object = Wireframe::page($event->object, [
+                'wireframe' => $this,
+            ]);
             $event->return = $event->object;
         }
     }
@@ -660,9 +685,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Getter method for specific class properties
      *
-     * Note that this differs notably from the parent class' get() method: unlike
-     * in WireData, here we limit the scope of the method to specific, predefined
-     * class properties instead of returning any index from the "data" array. We
+     * Note that this differs notably from the parent class' get() method: unlike in WireData, here we limit the scope
+     * of the method to specific, predefined class properties instead of returning any index from the "data" array. We
      * also don't support pipe ("|") separated strings or objects as arguments.
      *
      * @param string $key Name of property you want to retrieve.
@@ -741,8 +765,7 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Set values from an array
      *
-     * This method is a wrapper for the set() method, with support for multiple
-     * values as an associative array.
+     * This method is a wrapper for the set() method, with support for multiple values as an associative array.
      *
      * @param array $values Values as an associative array.
      * @return Wireframe Self-reference.
@@ -760,9 +783,8 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
     /**
      * Fetch a list of files recursively
      *
-     * This is a helper method used for fetching a list of files and folders
-     * recursively and returning the result as an object. Originally added
-     * for storing partial file references in an easy to access way.
+     * This is a helper method used for fetching a list of files and folders recursively and returning the result as an
+     * object. Originally added for storing partial file references in an easy to access way.
      *
      * @param string $path Base directory.
      * @param string $ext File extension.
@@ -785,6 +807,168 @@ class Wireframe extends WireData implements Module, ConfigurableModule {
             $this->cache[$cache_key] = $files;
         }
         return $files;
+    }
+
+    /**
+     * Static getter (factory) method for Components
+     *
+     * Note: keep in mind that due to file system differences and the use of an autoloader, the name of the component
+     * should *always* be treated as case sensitive. If actual class name is `Card` and the name is provided for this
+     * method as `card`, this will fail in some environments, resulting in an exception.
+     *
+     * @param string $component_name Component name.
+     * @param array $args Arguments for the Component.
+     * @return \Wireframe\Component Instance of the Component.
+     *
+     * @throws WireException if Component class isn't found.
+     */
+    public static function component(string $component_name, array $args = []): \Wireframe\Component {
+
+        $component = null;
+        $component_class = '\Wireframe\Component\\' . $component_name;
+
+        if (class_exists($component_class)) {
+            $reflector = new \ReflectionClass($component_class);
+            $component = $reflector->newInstanceArgs($args);
+        } else {
+            throw new WireException(sprintf(
+                'Component class %s was not found.',
+                $component_class
+            ));
+        }
+
+        return $component;
+    }
+
+    /**
+     * Static getter (factory) method for Pages
+     *
+     * This utility method can be used to render Page objects via Wireframe even if they don't have the expected
+     * altFilename Template setting in place. Particularly useful for cases where you don't want a page to be publicly
+     * viewable, but you still want to render it manually on some occasion (e.g. content that is only shown in lists.)
+     *
+     * Note that this method will accept different types of parameters, and the return value also depends on provided
+     * parameters. Basic usage:
+     *
+     * ```
+     * <?= Wireframe::page('id=1234', ['layout' => null, 'view' => 'list-item'])->render() ?>
+     * ```
+     *
+     * Or a shorter version with string provided for args:
+     *
+     * ```
+     * <?= Wireframe::page('id=1234', 'list-item') ?>
+     * ```
+     *
+     * @param int|string|Page $source Page object, Page ID (integer), or selector string (string).
+     * @param array|string $args Optional arguments. If the value is a string, it is assumed to be the name of a view
+     *                           file and the default value for layout is set to `null` – except if the string contains
+     *                           a forward slash in it, in which case it is assumed to hold both layout and view file
+     *                           names ([layout]/[view]). If the value is an array, following options are supported:
+     *                           - parent [Page]: the page on/for which current page is being rendered
+     *                           - wireframe [Wireframe]: an instance of the Wireframe module
+     *                           - wire [ProcessWire]: an instance of ProcessWire, defaults to Page's Wire instance if
+     *                             $source is a Page object, or the Wire instance returned by wire() method if not.
+     *                           - filename [string]: template file, defaults to 'wireframe'
+     *                           - ext [string]: extension for the template file, defaults to '.php'
+     *                           - layout [string]: layout to render the page with, defaults to 'default'
+     *                           - view [string]: view file to render the page with, defaults to 'default'
+     *                           - render [bool]: defines if we should return rendered content, defaults to 'false'
+     * @return string|Page|NullPage Returns string if 'render' option was 'true' **or** the args param was a string,
+     *                              otherwise returns a Page, or NullPage (if page wasn't found).
+     *
+     * @throws WireException if source param is of an unexpected type.
+     * @throws WireException if args param is of an unexpected type.
+     */
+    public static function page($source, $args = []) {
+
+        // ProcessWire instance
+        $wire = $args['wire'] ?? ($source instanceof Page ? $source->getWire() : wire());
+
+        // get a page
+        $page = null;
+        if ($source instanceof Page) {
+            $page = $source;
+        } else if (is_int($source) || is_string($source)) {
+            $page = $wire->pages->get($source);
+        } else {
+            throw new WireException(sprintf(
+                'Invalid argument type supplied for param source (%s)',
+                gettype($source) . (is_object($source) ? ' ' . get_class($source) : '')
+            ));
+        }
+
+        // bail out early if page wasn't found
+        if ($page instanceof NullPage) return $page;
+
+        // parse arguments and merge with defaults
+        if (is_string($args)) {
+            $args = [
+                'layout' => null,
+                'view' => $args,
+                'render' => true,
+            ];
+            if (strpos($args['view'], '/') !== false) {
+                $args_parts = explode('/', $args['view']);
+                $args = [
+                    'layout' => $args_parts[0],
+                    'view' => $args_parts[1],
+                    'render' => true,
+                ];
+            }
+        }
+        if (is_array($args)) {
+            $args = array_merge([
+                'parent' => null,
+                'wireframe' => null,
+                'wire' => $wire,
+                'filename' => null,
+                'ext' => '.php',
+                'layout' => 'default',
+                'view' => 'default',
+                'render' => false,
+            ], $args);
+        } else {
+            throw new WireException(sprintf(
+                'Invalid argument type supplied for param args (%s)',
+                gettype($args) . (is_object($args) ? ' ' . get_class($args) : '')
+            ));
+        }
+
+        // make sure that the page gets rendered with Wireframe
+        $page->_wireframe_filename = $args['filename'] ?? 'wireframe';
+        if (empty($args['filename']) && !empty($args['parent'])) {
+            $page->_wireframe_filename = $args['parent']->template->altFilename;
+        }
+        if (empty($page->template->altFilename) || $page->template->altFilename != $page->_wireframe_filename) {
+            $page->addHookBefore('render', function(HookEvent $event) use ($args) {
+                if (!empty($event->object->_wireframe_filename) && empty($event->object->template->altFilename)) {
+                    $options = $event->arguments[0] ?? [];
+                    if (empty($options['filename'])) {
+                        $options['filename'] = $event->object->_wireframe_filename . $args['ext'];
+                        $event->arguments(0, $options);
+                    }
+                }
+            });
+            if (!empty($args['wireframe']) && !empty($args['parent'])) {
+                $page->addHookAfter('render', function(HookEvent $event) use ($args) {
+                    if (!empty($event->object->_wireframe_page)) {
+                        $args['wireframe']->page = $args['parent'];
+                    }
+                });
+            }
+        }
+
+        // make sure that basic Wireframe features have been intiialized
+        if (!Wireframe::isInitialized($wire->instanceID)) {
+           ($args['wireframe'] ?? $wire->modules->get('Wireframe'))->initOnce();
+        }
+
+        // set view and layout
+        if ($args['layout'] != 'default') $page->setLayout($args['layout']);
+        if ($args['view'] != 'default') $page->setView($args['view']);
+
+        return $args['render'] ? $page->render() : $page;
     }
 
 }